Divide livingHinges from gui tool version of living hinges
